+++ docs_src/responses/orjson.py
@@ -1,23 +1,13 @@
-from esmerald import (
-    APIView,
-    Esmerald,
-    Gateway,
-    ORJSONResponse,
-    Request,
-    Response,
-    get,
-    post,
-    status,
-)
+from esmerald import APIView, Esmerald, Gateway, OrJSON, Request, get, post, status
 
 
 class World(APIView):
     @get(path="/{url}")
-    async def home(request: Request, url: str) -> ORJSONResponse:
-        return Response(f"URL: {url}")
+    async def home(self, request: Request, url: str) -> OrJSON:
+        return OrJSON(content=f"URL: {url}")
 
     @post(path="/{url}", status_code=status.HTTP_201_CREATED)
-    async def mars(request: Request, url: str) -> ORJSONResponse:
+    async def mars(self, request: Request, url: str) -> OrJSON:
         ...