+++ esmerald/__init__.py
@@ -1,7 +1,7 @@
 """
 Esmerald: Highly scalable, performant, easy to learn and for every application.
 """
-__version__ = "0.9.0"
+__version__ = "0.10.0"
 
 from starlette import status